Export class instead of instance Move NhostClient to its own module
